Advanced in the final path making in the gif

+++ Image_Manager.py
@@ -24,32 +24,39 @@ class Image_Manager:
         self._fig, self._ax = ox.plot_graph(self._map, figsize=(25, 25), show=False, close=False)
 
     def create_gif(self, with_path=False, route: list[int]=None):
-        with io.get_writer(f"{self._path}/animation.gif", mode='I', duration=1 / self._fps, loop=0) as writer:
-            for filename in self._filenames:
-                image = io.imread(filename)
+        writer = io.get_writer(f"{self._path}/animation.gif", mode='I', duration=1 / self._fps, loop=0)
+
+        # with io.get_writer(f"{self._path}/animation.gif", mode='I', duration=1 / self._fps, loop=0) as writer:
+        for filename in self._filenames:
+            image = io.imread(filename)
+            writer.append_data(image)
+        
+        if with_path:
+            # TODO: Create the gif with the path
+            print("Entered to create the gif with the path")
+            self.plot_graph_route(route)
+
+            final_path = f"{self._path}/finalpath.png"
+
+            i = 0
+            while not os.path.exists(final_path):
+                time.sleep(0.05)  # Sleep for 50ms
+                i += 1
+                if i == 100:
+                    print("The final image was not created")
+                    break
+
+            image = io.imread(f"{self._path}/finalpath.png")
+            for _ in range(120):  # 4 seconds
+                print("Entered path to the gif")
                 writer.append_data(image)
-            
-            if with_path:
-                # TODO: Create the gif with the path
-                print("Entered to create the gif with the path")
-                self.plot_graph_route(route)
 
-                final_path = f"{self._path}/finalpath.png"
-
-                i = 0
-                while not os.path.exists(final_path):
-                    time.sleep(0.05)  # Sleep for 50ms
-                    i += 1
-                    if i == 100:
-                        print("The final image was not created")
-                        break
-
-                image = io.imread(f"{self._path}/finalpath.png")
-                for _ in range(120):  # 4 seconds
-                    print("Entered path to the gif")
-                    writer.append_data(image)
+        writer.close()
 
         for filename in self._filenames:
+            if filename == "Why.txt":
+                continue
+            
             os.remove(filename)
 
     def graph_main_nodes(self, start: dict[str, float], end: dict[str, float]):
@@ -68,8 +75,8 @@ class Image_Manager:
         self._filenames.append(f"{self._frames_path}/{self._counter}.png")
         self._counter += 1
 
-    # def finish_plot(self):
-    #     plt.close(self._fig)
+    def finish_plot(self):
+        plt.close(self._fig)
 
     def plot_graph_route(self, route: list[int], color: str = 'b', linewidth: int = 6, node_size: int = 0, bgcolor: str = 'k', figsize: tuple[int, int] = (20, 20)):
         print("entered to plot the route")
